Starting to add a bit of substance to this library
Gradually adding stuff to this library at its lowest level of functionality, but without quite as much code refactoring as would be advised later on, as this is ancient code and has some inefficiencies in how it uses collecgtions and boxing of integers. The enum of Chart Types is handy and might grow later on, and the Chart legend is a convenient but not very flexible basic facility for adding an icon and label based data set key lookup for charts that don't include legends as part of their API, such as the basic one that I will be adding to this toolkit soon as a way of parking some ancient legacy code that nevertheless may still prove useful to people who just need simple charting vs. something as flexible and comprehensive as JFreeChart.